+++ src/core/tile_mipmapper.rs
@@ -169,6 +169,9 @@ impl TileCollector for TileMipmapper<'_> {
 
 		fs::create_dir_all(&self.config.tile_dir(TileKind::Map, level))?;
 		fs::create_dir_all(&self.config.tile_dir(TileKind::Lightmap, level))?;
+		if self.config.height_layer {
+			fs::create_dir_all(&self.config.tile_dir(TileKind::Heightmap, level))?;
+		}
 
 		Ok(())
 	}
@@ -204,7 +207,15 @@ impl TileCollector for TileMipmapper<'_> {
 		let map_stat = self.render_mipmap::<image::Rgba<u8>>(TileKind::Map, level, coords, prev)?;
 		let lightmap_stat =
 			self.render_mipmap::<image::LumaA<u8>>(TileKind::Lightmap, level, coords, prev)?;
-		Ok(map_stat + lightmap_stat)
+		let height_stat = if self.config.height_layer {
+			self.render_mipmap::<image::Rgba<u8>>(TileKind::Heightmap, level, coords, prev)?
+		} else {
+			MipmapStat {
+				total: 0,
+				processed: 0,
+			}
+		};
+		Ok(map_stat + lightmap_stat + height_stat)
 	}
 }